ClassroomMoodAI/requirements.txt
torch
torchvision
transformers
pandas
pillow